* 1.0-integration: (138 commits)
  [x11] Disable XInput by default
  [xinput] Invert the XI extension version check
  [cogl-primitives] Fix an unused variable warning when building GLES
  [clutter-stage-egl] Pass -1,-1 to clutter_stage_x11_fix_window_size
  Update the GLES backend to have the layer filters in the material
  [gles/cogl-shader] Add a missing semicolon
  [cogl] Move the texture filters to be a property of the material layer
  [text] Fix Pango unit to pixels conversion
  [actor] Force unrealization on destroy only for non-toplevels
  [x11] Rework map/unmap and resizing
  [xinput] Check for the XInput entry points
  [units] Validate units against the ParamSpec
  [actor] Add the ::allocation-changed signal
  [actor] Use flags to control allocations
  [units] Rework Units into logical distance value
  Remove a stray g_value_get_int()
  Remove usage of Units and macros
  [cogl-material] Allow setting a layer with an invalid texture handle
  [timeline] Remove the concept of frames from timelines
  [gles/cogl-shader] Fix parameter spec for cogl_shader_get_info_log
  ...

Conflicts:
	configure.ac

+++ clutter/clutter-event.c
@@ -120,10 +120,10 @@ clutter_event_get_state (ClutterEvent *event)
  */
 void
 clutter_event_get_coords (ClutterEvent *event,
-                          gint         *x,
-                          gint         *y)
+                          gfloat       *x,
+                          gfloat       *y)
 {
-  gint event_x, event_y;
+  gfloat event_x, event_y;
 
   g_return_if_fail (event != NULL);
 
@@ -141,15 +141,18 @@ clutter_event_get_coords (ClutterEvent *event,
     case CLUTTER_ENTER:
     case CLUTTER_LEAVE:
       break;
+
     case CLUTTER_BUTTON_PRESS:
     case CLUTTER_BUTTON_RELEASE:
       event_x = event->button.x;
       event_y = event->button.y;
       break;
+
     case CLUTTER_MOTION:
       event_x = event->motion.x;
       event_y = event->motion.y;
       break;
+
     case CLUTTER_SCROLL:
       event_x = event->scroll.x;
       event_y = event->scroll.y;
