+++ tests/interactive/test-easing.c
@@ -53,8 +53,8 @@ on_button_press (ClutterActor       *actor,
   if (event->button == 3)
     {
       gchar *text;
-      guint stage_width, stage_height;
-      guint label_width, label_height;
+      gfloat stage_width, stage_height;
+      gfloat label_width, label_height;
 
       current_mode = (current_mode + 1 < n_easing_modes) ? current_mode + 1
                                                          : 0;
@@ -110,8 +110,8 @@ test_easing_main (int argc, char *argv[])
   ClutterColor stage_color = { 0x66, 0x66, 0xdd, 0xff };
   ClutterColor rect_color = { 0x44, 0xdd, 0x44, 0xff };
   gchar *text;
-  guint stage_width, stage_height;
-  guint label_width, label_height;
+  gfloat stage_width, stage_height;
+  gfloat label_width, label_height;
 
   clutter_init (&argc, &argv);
